kp 1.9.9

+++ koolproxy/koolproxy/install.sh
@@ -10,7 +10,15 @@ rm -rf /koolshare/koolproxy/rule*
 
 # copy new files
 cd /tmp
-cp -rf /tmp/koolproxy/koolproxy /koolshare/
+
+if [ ! -f / /koolshare/koolproxy/data/user.txt ];then
+	cp -rf /tmp/koolproxy/koolproxy /koolshare/
+else
+	mv /koolshare/koolproxy/data/user.txt /tmp/user.txt.tmp
+	cp -rf /tmp/koolproxy/koolproxy /koolshare/
+	mv /tmp/user.txt.tmp /koolshare/koolproxy/data/user.txt
+fi
+
 cp -rf /tmp/koolproxy/scripts/* /koolshare/scripts/
 cp -rf /tmp/koolproxy/webs/* /koolshare/webs/
 cp -rf /tmp/koolproxy/res/* /koolshare/res/
@@ -22,9 +30,18 @@ chmod 755 /koolshare/koolproxy/*
 chmod 755 /koolshare/scripts/*
 chmod 755 /koolshare/perp//koolproxy/*
 
+if [ -z "$koolproxy_dbug" ];then
+	dbus set koolproxy_dbug=1
+fi
 
-dbus set koolproxy_policy=1
-dbus set koolproxy_lan_control=0
+if [ -z "$koolproxy_policy" ];then
+	dbus set koolproxy_policy=1
+fi
+
+if [ -z "$koolproxy_lan_control" ];then
+	dbus set koolproxy_lan_control=0
+fi
+
 sleep 1
 # start
 if [ "$koolproxy_enable" == "1" ];then
